+++ source/blood/src/triggers.cpp
@@ -2885,7 +2885,7 @@ void pastePropertiesInObj(int type, int nDest, EVENT event) {
         /* - hitag: 2 = force swap min and max if goal value was reached								- */
         /* - hitag: 3 = force reset counter	                                           					- */
 
-        long data = getDataFieldOfObject(type, nDest, pXSource->data1);
+        int data = getDataFieldOfObject(type, nDest, pXSource->data1);
         if (data == -65535) return;
 
         if (pXSource->data2 < pXSource->data3) {
@@ -4511,7 +4511,7 @@ void UniMissileTrapSeqCallback(int, int nXSprite)
         // add velocity controlling
         if (pXSprite->data2 > 0) {
             
-            long velocity = pXSprite->data2 << 12; 
+            int velocity = pXSprite->data2 << 12;
             xvel[pMissile->xvel] = mulscale(velocity, dx, 14);
             yvel[pMissile->xvel] = mulscale(velocity, dy, 14);
             zvel[pMissile->xvel] = mulscale(velocity, dz, 14);
